Use Symfony's Process component
Use the local machine's `ssh` client to handle SSH connections. This replaces the phpseclib library, and instead uses Symfony's Process component to perform calls to `ssh`.
This move brings the advantage that the SSH client can be configured by the user, and configured public keys will work as expected.

A new method `prepareCommand` has been added to the `WordpressInstance` class. This can be used to prepare a command suitable for passing to Symfony's Process component. Using this method, it's possible to specify which command you want to run on the container, as well as any additional `ssh` and `docker exec` command line arguments.